Delete_JobTestController
